Fixed compiler errors in this ancient code library, but file loads cause crashes
Lots of compiler errors in this code, so it took almost an hour and a half to get it into working shape, but it crashes when loading files. The GUI looks and behaves great, so it's possible that some of the compiler fixes caused casting issues leading to null file pointers.